improved ui interface

+++ src/components/userCard.vue
@@ -1,5 +1,13 @@
 <template>
   <v-card :loading="loading">
+    <v-toolbar color="transparent" dense flat>
+      <v-toolbar-title>{{ item.email }}</v-toolbar-title>
+      <v-spacer></v-spacer>
+      <v-btn icon @click="del" color="error">
+        <v-icon>mdi-delete</v-icon>
+      </v-btn>
+    </v-toolbar>
+    <v-divider></v-divider>
     <v-list-item three-line>
       <v-list-item-avatar
         size="125"
@@ -10,12 +18,9 @@
 
       <v-list-item-content class="align-self-start">
         <v-list-item-title
-          class="headline mb-2"
-          v-text="item.email"
-        ></v-list-item-title>
+          class="title mb-2"
+        >{{item.displayName | nameCheck}}</v-list-item-title>
 
-        <!-- <v-list-item-subtitle v-text="item.displayName | nameCheck"></v-list-item-subtitle> -->
-        <v-list-item-subtitle>{{item.displayName | nameCheck}}</v-list-item-subtitle>
         <v-list-item-subtitle>
           <v-select
             class="ma-2"
@@ -28,10 +33,6 @@
         </v-list-item-subtitle>
       </v-list-item-content>
     </v-list-item>
-    <v-card-actions>
-      <v-spacer></v-spacer>
-      <v-btn @click="del" color="error">삭제</v-btn>
-    </v-card-actions>
   </v-card>
 </template>
 <script>
@@ -54,23 +55,47 @@ export default {
     },
     imgCheck (v) {
       if (v) return v
-      return 'https://cdn.vuetifyjs.com/images/cards/foster.jpg'
+      return require('@/assets/images/account.png')
     }
   },
   methods: {
-    levelChange (v) {
+    async levelChange (v) {
+      const r = await this.$swal.fire({
+        title: '정말 변경하시겠습니까?',
+        text: '변경 후 되돌릴 수 없습니다.',
+        type: 'warning',
+        confirmButtonText: '확인',
+        cancelButtonText: '취소',
+        showCancelButton: true
+      })
+      if (!r.value) return
       this.loading = true
-      this.$axios.patch(`/admin/user/${v.uid}/level`, { level: v.level })
-        .catch(e => {
-          this.$toasted.global.error(e.message)
-        })
-        .finally(() => {
-          this.loading = false
-        })
+      try {
+        await this.$axios.patch(`/admin/user/${v.uid}/level`, { level: v.level })
+      } catch (e) {
+        this.$toasted.global.error(e.message)
+      } finally {
+        this.loading = false
+      }
     },
     async del () {
-      await this.$axios.delete(`/admin/user/${this.item.uid}`)
-      this.$emit('del')
+      const r = await this.$swal.fire({
+        title: '정말 삭제하시겠습니까?',
+        text: '삭제 후 되돌릴 수 없습니다.',
+        type: 'error',
+        // confirmButtonText: 'Cool',
+        showCancelButton: true
+      })
+      if (!r.value) return
+      this.loading = true
+      try {
+        await this.$axios.delete(`/admin/user/${this.item.uid}`)
+        this.$emit('del')
+      } catch (e) {
+        this.$toasted.global.error(e.message)
+      } finally {
+        this.loading = false
+      }
     }
   }
 }
